Convert to explicitly sized numeric types
We currently have int which is a 64-bit integer, and float, which is a
double precision IEEE754 value.  Postgres supports other numeric
precisions, and it makes sense to expose them.

Naming-wise we follow the modern convention of specifying the bit size
of the type as a suffix, so the new fixed-precision numeric types are:

- `std::int16` -- 16-bit signed integer
- `std::int32` -- 32-bit signed integer
- `std::int64` -- 64-bit signed integer
- `std::float32` -- single presision IEEE754 floating-point type
- `std::float64` -- double presision IEEE754 floating-point type

Additionally `std::decimal` has been renamed to `std::numeric`, and
the following abstract scalar types have been introduced to describe
the numeric type groups:

- `std::anyint` -- any integer type
- `std::anyfloat` -- any floating-point type
- `std::anyreal` -- any non-complex numeric type, includes
  `anyint` and `anyfloat`, as well as `numeric`.

Fixes: #135.

+++ edgedb/lang/graphql/types.py
@@ -239,16 +239,16 @@ class GQLDirectiveLocation(_GQLType):
 
 PY_COERCION_MAP = {
     str: (s_types.string.Str, s_types.uuid.UUID),
-    int: (s_types.int.Int, s_types.numeric.Float, s_types.numeric.Decimal,
+    int: (s_types.int.Int64, s_types.numeric.Float64, s_types.numeric.Numeric,
           s_types.uuid.UUID),
-    float: (s_types.numeric.Float, s_types.numeric.Decimal),
+    float: (s_types.numeric.Float64, s_types.numeric.Numeric),
     bool: s_types.boolean.Bool,
 }
 
 GQL_TYPE_NAMES_MAP = {
     'String': s_types.string.Str,
-    'Int': s_types.int.Int,
-    'Float': s_types.numeric.Float,
+    'Int': s_types.int.Int64,
+    'Float': s_types.numeric.Float64,
     'Boolean': s_types.boolean.Bool,
     'ID': s_types.uuid.UUID,
 }
